论文搜索引擎网站/requirements.txt
# 核心依赖
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6

# LangChain相关（与0.3生态兼容的拆分包）
langchain-core==0.3.4
langchain-community==0.3.3

# 向量与嵌入
chromadb==0.5.5
sentence-transformers==3.0.1

# 大模型相关（Qwen/Transformers）
transformers==4.43.3
torch>=2.1.0
accelerate==0.33.0

# 数据处理
pandas==2.1.3
numpy==1.24.3
requests==2.31.0
beautifulsoup4==4.12.2

# ArXiv API
arxiv==2.1.0

# 文档处理
pypdf2==3.0.1
python-docx==1.1.0

# 环境变量
python-dotenv==1.0.0

# 日志
loguru==0.7.2

# 测试
pytest==7.4.3
pytest-asyncio==0.21.1

# 开发工具
black==23.11.0
flake8==6.1.0
